requirements.txt
graphene
graphene-django
graphql-core>=2.1rc1
django
django-filter>=2
django-graphql-jwt
graphene-django-extras
graphene-django-extras[date]
